requirements.txt
# AI Tools Assignment Dependencies
tensorflow>=2.15.0
scikit-learn
spacy
pandas
matplotlib
numpy
jupyter
streamlit
pillow
streamlit-drawable-canvas